requirements.txt
click
tabulate
eth-keys
eth-utils
pynacl
coincurve
mnemonic
eth-hash[pycryptodome]